move BackendProvider impl into backend companion objects
