+++ kool-backend-wgpu4k/src/commonMain/kotlin/de/fabmax/kool/pipeline/backend/wgpu/WgpuRenderBackend.kt
@@ -414,5 +414,4 @@ abstract class WgpuRenderBackend(
             this += computeSrc
         }
     }
-
 }
