+++ kool-backend-wgpu4k/src/desktopMain/kotlin/de/fabmax/kool/pipeline/backend/wgpu/RenderBackendWgpu4k.kt
@@ -3,16 +3,12 @@ package de.fabmax.kool.pipeline.backend.wgpu
 import de.fabmax.kool.KoolContext
 import de.fabmax.kool.pipeline.backend.BackendProvider
 import de.fabmax.kool.pipeline.backend.RenderBackend
-import de.fabmax.kool.platform.JsContext
+import de.fabmax.kool.platform.Lwjgl3Context
 
-object Wgpu4kBackendProvider : BackendProvider {
+object RenderBackendWgpu4k : BackendProvider {
     override val displayName: String = "wgpu4k"
 
     override fun createBackend(ctx: KoolContext): Result<RenderBackend> {
-        return if (JsRenderBackendWebGpu.isSupported()) {
-            Result.success(createWGPURenderBackend(ctx as JsContext))
-        } else {
-            Result.failure(IllegalStateException("WebGPU not supported"))
-        }
+        return Result.success(createWGPURenderBackend(ctx as Lwjgl3Context))
     }
 }
