+++ kool-core/src/androidMain/kotlin/de/fabmax/kool/pipeline/backend/gl/RenderBackendGlImpl.kt
@@ -12,6 +12,8 @@ import javax.microedition.khronos.opengles.GL10
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 
+actual fun createRenderBackendGl(ctx: KoolContext): RenderBackendGl = RenderBackendGlImpl(ctx as KoolContextAndroid)
+
 class RenderBackendGlImpl(ctx: KoolContextAndroid) :
     RenderBackendGl(1, GlImpl, ctx),
     GLSurfaceView.Renderer
